requirements.txt
fastapi
uvicorn[standard]
gunicorn
botbuilder-core
botbuilder-schema
aiosqlite
SQLModel
openai
python-dotenv
aiohttp
azure-identity
google-auth
google-auth-oauthlib
google-api-python-client
msal
requests
teams-ai